pi-alice/BlackRoad-Operating-System/scripts/cloudflare/requirements.txt
# Cloudflare DNS Management Scripts Requirements
# Install with: pip install -r scripts/cloudflare/requirements.txt

# HTTP client for API requests
requests>=2.31.0

# YAML parsing for domains.yaml
PyYAML>=6.0.1

# DNS resolution verification
dnspython>=2.4.2

# Colored terminal output
colorama>=0.4.6
